requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
flake8==3.9.2
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-speech==2.9.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
libcst==0.3.21
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.0
pandas==1.3.3
proto-plus==1.19.2
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydub==0.25.1
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
